account/requirements.txt
django==3.0.3
djangorestframework==3.11.0
django-filter==2.2.0
psycopg2==2.8.5
gunicorn==20.0.4
requests==2.23.0
django-cors-headers==3.3.0
drf-yasg==1.17.1
djangorestframework-simplejwt==4.4.0
cryptography==2.9.2